Move json schema decoding to transport pkg
